[TabsManager] Write DocumentNode
